chore: extractor windows fix + rspeedy build order + playground dev plumbing
Independent slice of the broader a2ui refactor — these changes don't
depend on the renderer/store reshape and can land on their own.

- a2ui-catalog-extractor: force POSIX path separators on TypeDoc
  entry points so the extractor works on Windows; exit non-zero when
  source files match but 0 components are emitted; recognize the
  published bin shim as a valid entry script.
- rspeedy/core/turbo.json: declare `//#build` as a dep of `build`.
  Without it, a fresh CI checkout can race rspeedy's `rslib build`
  d.ts generation against the root `tsc --build` that produces the
  webpack plugins' `lib/*.d.ts`, intermittently failing with TS2307
  on `@lynx-js/cache-events-webpack-plugin` /
  `@lynx-js/chunk-loading-webpack-plugin`.
- web-platform/web-core: stop redeclaring `fetch` as a chunk-scope
  binding so BTS chunks reuse `window.fetch`.
- a2ui-playground: relocate the in-memory payload store from
  `lynx.config.ts` to `rsbuild.config.ts` (it belongs to the rsbuild
  dev server, not the rspeedy lynx config); run `gcPayloads()` on
  GET too so an idle dev server can't keep stale payloads past the
  30-minute TTL; add per-package `turbo.json` declaring task
  ordering and outputs.
- ci: ignore `packages/genui/a2ui-playground/www/**` in eslint and
  vitest coverage so the rspeedy/rsbuild build output isn't parsed
  as source.
