+++ eslint.config.js
@@ -33,6 +33,9 @@ export default tseslint.config(
       '**/test/js',
       '**/dist/**',
       '**/lib/**',
+      // rspeedy/rsbuild build output for the a2ui-playground (lynx bundle
+      // + the rsbuild static site that serves it).
+      'packages/genui/a2ui-playground/www/**',
       'website/doc_build/**',
       '.changeset/*',
       '**/CHANGELOG.md',
