+++ packages/rspeedy/core/turbo.json
@@ -6,6 +6,7 @@
   "tasks": {
     "build": {
       "dependsOn": [
+        "//#build",
         "^build"
       ],
       "env": [
