+++ packages/web-platform/web-core/ts/client/background/background-apis/createChunkLoading.ts
@@ -89,7 +89,7 @@ export function createChunkLoading(entryTemplateUrl: string): {
       'alert',
       'confirm',
       'prompt',
-      'fetch',
+      // 'fetch',
       'XMLHttpRequest',
       'webkit',
       'Reporter',
@@ -134,7 +134,7 @@ export function createChunkLoading(entryTemplateUrl: string): {
           tt.alert,
           tt.confirm,
           tt.prompt,
-          tt.fetch,
+          // tt.fetch,
           tt.XMLHttpRequest,
           tt.webkit,
           tt.Reporter,
