+++ vitest.config.ts
@@ -17,6 +17,8 @@ export default defineConfig({
         '**/*.bench.js',
         '**/*.bench.ts',
         '**/dist/**',
+        // rspeedy/rsbuild build output for the a2ui-playground.
+        'packages/genui/a2ui-playground/www/**',
         '.github/**',
         'examples/**',
         'packages/**/lib/**',
